chore #9: AccompanyRegionType 이넘 AccompanyPost 내부 클래스로 이동
