+++ src/main/java/com/gogoring/dongoorami/accompany/presentation/AccompanyController.java
@@ -1,7 +1,7 @@
 package com.gogoring.dongoorami.accompany.presentation;
 
 import com.gogoring.dongoorami.accompany.application.AccompanyService;
-import com.gogoring.dongoorami.accompany.domain.AccompanyRegionType;
+import com.gogoring.dongoorami.accompany.domain.AccompanyPost.AccompanyRegionType;
 import com.gogoring.dongoorami.accompany.dto.request.AccompanyCommentRequest;
 import com.gogoring.dongoorami.accompany.dto.request.AccompanyPostFilterRequest;
 import com.gogoring.dongoorami.accompany.dto.request.AccompanyPostRequest;
